Dashboard overview section overdue tasks bug fixed

+++ lib/model/all_categories_performance_report_model.dart
@@ -64,7 +64,7 @@ class AllCategoriesPerformanceReportModel {
     overdueTasks = json['overdueTasks'];
     onTimeTasks = json['onTimeTasks'];
     delayedTasks = json['delayedTasks'];
-    completionRate = (json['completionRate'] as double).toInt();
+    completionRate = (json['completionRate']);
   }
 
   Map<String, dynamic> toJson() {
